refactored login and register routes into auth route
