+++ index.js
@@ -1,8 +1,9 @@
 const express = require("express");
 const helmet = require("helmet");
-const register = require("./routers/register");
+// const register = require("./routers/register");
 const users = require("./routers/users");
-const login = require("./routers/login");
+// const login = require("./routers/login");
+const auth = require("./routers/auth");
 
 const app = express();
 
@@ -10,9 +11,10 @@ app.use(express.json());
 app.use(helmet());
 
 app.get("/", (req, res) => res.send("<h1>WebAuth I</h1>"));
-app.use("/api/register", register);
+// app.use("/api/register", register);
 app.use("/api/users", users);
-app.use("/api/login", login);
+// app.use("/api/login", login);
+app.use("/api/auth", auth);
 
 const port = process.env.PORT || 6500;
 app.listen(port, console.log(`Server running on ${port}`));
